Add files via upload
In this project, we constructed a light detector circuit using a NAND gate chip. A light detector circuit is a circuit that can detect light. When a bright light shines on the circuit, such as from a flashlight, the LED turns on and when the flash light is taken away from the circuit then the LED turns off.